Move tasks to pkg/utils (#3028)
